将主键类型由Integer换成Long

+++ src/main/java/com/dboper/cud/core/DomainOper.java
@@ -14,8 +14,8 @@ public interface DomainOper {
 	public MyResponseBody insertList(List<DBBody> bodies);
 	public MyResponseBody delete(DBBody body);
 	public MyResponseBody update(String table_name,Map<String,Object> body,Map<String,Object> params);
-	public MyResponseBody deleteById(String table_name,int id);
+	public MyResponseBody deleteById(String table_name,Long id);
 	public MyResponseBody deleteByForeign(String table_name,String foreign_name,Object foreign_id);
-	public Map<Integer,MyResponseBody> deleteBatchById(String table_name,List<Integer> ids);
+	public Map<Long,MyResponseBody> deleteBatchById(String table_name,List<Long> ids);
 	public int executeSqlCount(String sql);
 }
